Move AccountsTab to subdirectory
